requirements.txt
numpy>=1.21.5
pytorch>=1.12.0
transformers>=4.20.1
pandas>=1.3.5
tqdm>=4.64.0
rdkit>=2020.09.1
sklearn>=1.0.2
cudatoolkit>=10.1.243
